Мержик

+++ modular_sand/code/game/objects/effects/decals/cleanable/lewd_decals.dm
@@ -10,17 +10,22 @@
 
 /obj/effect/decal/cleanable/semendrip/replace_decal(obj/effect/decal/cleanable/semendrip/C)
 	. = ..()
-	reagents.trans_to(src, C.reagents.total_volume)
-	transfer_blood_dna(C.blood_DNA)
+	if(!. || QDELETED(src))
+		return FALSE
 	var/obj/effect/decal/cleanable/semen/S = (locate(/obj/effect/decal/cleanable/semen) in C.loc)
-	if(S)
-		C.reagents.trans_to(S, C.reagents.total_volume)
-		C.transfer_blood_dna(S.blood_DNA)
-		C.update_icon()
-		return
-	if(C.reagents.total_volume >= 10)
+	if(S) // Merge ourselves into this puddle.
+		if (reagents) // BLUEMOON EDIT: Invalid Space Turfs
+			reagents.trans_to(S, reagents.total_volume) // BLUEMOON EDIT: Invalid Space Turfs
+		S.transfer_blood_dna(blood_DNA)
+		update_icon()
+		return TRUE
+	if (reagents)
+		reagents.trans_to(C, reagents.total_volume)
+	C.transfer_blood_dna(blood_DNA)
+	if(C.reagents && C.reagents.total_volume >= 10) // Turn the drip into a puddle. // BLUEMOON EDIT: Invalid Space Turfs
 		S = new(C.loc)
-		C.reagents.trans_to(S, C.reagents.total_volume)
+		if (C.reagents)
+			C.reagents.trans_to(S, C.reagents.total_volume)  // BLUEMOON EDIT: Invalid Space Turfs
 		C.transfer_blood_dna(S.blood_DNA)
 		S.update_icon()
 		qdel(C)
